Additions: Linux: rcvboxadd: Introduce "status-kernel" and "status-user" actions, bugref:10359.
"rcvboxadd status-kernel" is intended to verify if Guest Additions kernel modules are currently
loaded and running versions do match to Guest Additions installation version.

"rcvboxadd status-user" is intended to verify if user-space processes are currently running. At
the moment only checks if VBoxService is running, can be extended later on.